__init__: time can be float

+++ pyopspec/steps/isothermal_step.py
@@ -2,7 +2,7 @@ class IsothermalStep():
     """
     """
 
-    def __init__(self, pressure:int, flow_rates:dict[str, float], time:int):
+    def __init__(self, pressure:int, flow_rates:dict[str, float], time:float):
         """
         """
